authenticate() accepts instance object
The previous signature for authenticate() took a callback as the first
arg and then keyword default args for the rest.  It was possible to pass
an instance object by calling authenticate(callback, **user.__dict__),
but made for a messy and unpredictable API with the existence of classic
old style classes.

Accepting an instance object also has the greater benefit of returning
the same instance object. This object 'passthrough' style allows the
caller of authenticate() to attach other attributes to the instance
object un-related to the functionality of authenticate, but important to
have in the callback.

 On branch master
	modified:   pychap.py
	modified:   test.py